don't take lookback days as input from the users. There are two flows: 1. Sample Size to MDE flow. 2. MDE to sample size flow.
in the first flow, user will provide sample size as input along with mean, variance and number of users in the lookback days. They'll not provide lookback days as input but will provide number of days they want to run the experiment as input. The answer should then be the MDE.

In the second flow, the user will give MDE as input. The calculator should then decide the sample size that will be needed for different time periods in weeks. The data for different mean, variance at different lookback days is present in an excel file. Provide an ability to upload excel file with this data for this flow.